Clase 16/requirements.txt
discord.py
python-dotenv
TikTokLive